Separating the board from App.js
